refactor(mastering): extract Stages 5+5.4 (verification + ceiling guard) (#290 D5)
