+++ servers/bitwize-music-server/handlers/processing/audio.py
@@ -651,8 +651,6 @@ async def master_album(
     # --- Stage 1: Pre-flight ---
     if result := await _album_stages._stage_pre_flight(ctx):
         return result
-    import numpy as np
-    from tools.mastering.analyze_tracks import analyze_track
 
     audio_dir = ctx.audio_dir
     source_dir = ctx.source_dir
@@ -701,294 +699,17 @@ async def master_album(
     output_dir = ctx.output_dir
     mastered_files = ctx.mastered_files
 
-    # Build eq_settings for auto-recovery in Stage 5 (mirrors Stage 4 logic).
-    eq_settings = []
-    if effective_highmid != 0:
-        eq_settings.append((3500.0, effective_highmid, 1.5))
-    if effective_highs != 0:
-        eq_settings.append((8000.0, effective_highs, 0.7))
-
     # --- Stage 5: Verification ---
-
-    verify_results = []
-    for wav in mastered_files:
-        result = await loop.run_in_executor(None, analyze_track, str(wav))
-        verify_results.append(result)
-
-    verify_lufs = [r["lufs"] for r in verify_results]
-    verify_avg = float(np.mean(verify_lufs))
-    verify_range = float(max(verify_lufs) - min(verify_lufs))
-
-    # Check thresholds
-    out_of_spec = []
-    for r in verify_results:
-        issues = []
-        if abs(r["lufs"] - effective_lufs) > 0.5:
-            issues.append(f"LUFS {r['lufs']:.1f} outside ±0.5 dB of target {effective_lufs}")
-        if r["peak_db"] > effective_ceiling:
-            issues.append(
-                f"Peak {r['peak_db']:.1f} dB exceeds ceiling {effective_ceiling} dB"
-            )
-        if issues:
-            out_of_spec.append({"filename": r["filename"], "issues": issues})
-
-    album_range_fail = verify_range >= 1.0
-    auto_recovered: list[dict[str, Any]] = []
-
-    if out_of_spec or album_range_fail:
-        # --- Auto-recovery: fix recoverable dynamic range issues ---
-        recoverable = []
-        for spec in out_of_spec:
-            has_peak_issue = any("Peak" in iss for iss in spec["issues"])
-            vr = next(
-                (r for r in verify_results if r["filename"] == spec["filename"]),
-                None,
-            )
-            if not vr:
-                continue
-            lufs_too_low = vr["lufs"] < effective_lufs - 0.5
-            peak_at_ceiling = vr["peak_db"] >= effective_ceiling - 0.1
-            if lufs_too_low and peak_at_ceiling and not has_peak_issue:
-                recoverable.append(spec["filename"])
-
-        if recoverable:
-            from tools.mastering.fix_dynamic_track import fix_dynamic
-
-            # Recovery writes at source rate (fix_dynamic is a rescue path
-            # that doesn't resample) but honors the configured output bit
-            # depth from targets["output_bits"].
-            recovery_subtype = (
-                "PCM_24" if targets["output_bits"] > 16 else "PCM_16"
-            )
-
-            auto_recovered = []
-            for fname in recoverable:
-                raw_path = source_dir / fname
-                if not raw_path.exists():
-                    raw_path = _find_wav_source_dir(audio_dir) / fname
-                if not raw_path.exists():
-                    continue
-
-                def _do_recovery(
-                    src: Path,
-                    dst: Path,
-                    lufs: float,
-                    eq: list[tuple[float, float, float]],
-                    ceil: float,
-                    subtype: str,
-                ) -> dict[str, Any]:
-                    import soundfile as sf
-                    data, rate = sf.read(str(src))
-                    if len(data.shape) == 1:
-                        data = np.column_stack([data, data])
-                    data, metrics = fix_dynamic(
-                        data, rate,
-                        target_lufs=lufs,
-                        eq_settings=eq if eq else None,
-                        ceiling_db=ceil,
-                    )
-                    sf.write(str(dst), data, rate, subtype=subtype)
-                    return metrics
-
-                mastered_path = output_dir / fname
-                metrics = await loop.run_in_executor(
-                    None, _do_recovery, raw_path, mastered_path,
-                    effective_lufs, eq_settings, effective_ceiling,
-                    recovery_subtype,
-                )
-                auto_recovered.append({
-                    "filename": fname,
-                    "original_lufs": metrics["original_lufs"],
-                    "final_lufs": metrics["final_lufs"],
-                    "final_peak_db": metrics["final_peak_db"],
-                })
-
-            if auto_recovered:
-                warnings.append({
-                    "type": "auto_recovery",
-                    "tracks_fixed": [r["filename"] for r in auto_recovered],
-                })
-
-                # Re-verify ALL tracks (album range check needs all)
-                verify_results = []
-                for wav in mastered_files:
-                    result = await loop.run_in_executor(
-                        None, analyze_track, str(wav),
-                    )
-                    verify_results.append(result)
-
-                verify_lufs = [r["lufs"] for r in verify_results]
-                verify_avg = float(np.mean(verify_lufs))
-                verify_range = float(max(verify_lufs) - min(verify_lufs))
-
-                out_of_spec = []
-                for r in verify_results:
-                    issues = []
-                    if abs(r["lufs"] - effective_lufs) > 0.5:
-                        issues.append(
-                            f"LUFS {r['lufs']:.1f} outside ±0.5 dB of target {effective_lufs}"
-                        )
-                    if r["peak_db"] > effective_ceiling:
-                        issues.append(
-                            f"Peak {r['peak_db']:.1f} dB exceeds ceiling {effective_ceiling} dB"
-                        )
-                    if issues:
-                        out_of_spec.append({"filename": r["filename"], "issues": issues})
-
-                album_range_fail = verify_range >= 1.0
-
-        # If still failing after recovery attempt, return failure
-        if out_of_spec or album_range_fail:
-            fail_detail: dict[str, Any] = {}
-            if out_of_spec:
-                fail_detail["tracks_out_of_spec"] = out_of_spec
-            if album_range_fail:
-                fail_detail["album_lufs_range"] = round(verify_range, 2)
-                fail_detail["album_range_limit"] = 1.0
-
-            stages["verification"] = {
-                "status": "fail",
-                "avg_lufs": round(verify_avg, 1),
-                "lufs_range": round(verify_range, 2),
-                "all_within_spec": False,
-            }
-            return _safe_json({
-                "album_slug": album_slug,
-                "stage_reached": "verification",
-                "stages": stages,
-                "settings": settings,
-                "warnings": warnings,
-                "failed_stage": "verification",
-                "failure_detail": fail_detail,
-            })
-
-    verification_stage = {
-        "status": "pass",
-        "avg_lufs": round(verify_avg, 1),
-        "lufs_range": round(verify_range, 2),
-        "all_within_spec": True,
-    }
-    # Include auto-recovery details when tracks were fixed
-    if auto_recovered:
-        verification_stage["auto_recovered"] = auto_recovered
-    stages["verification"] = verification_stage
+    if result := await _album_stages._stage_verification(ctx):
+        return result
 
     # --- Stage 5.4: Album-ceiling guard (#290 step 8) ---
-    # Gate every track against album_median + 2 LU. Small overshoots
-    # (<= 0.5 LU) get a silent scalar pull-down; large overshoots halt
-    # and escalate (signals coherence correction didn't converge on LUFS-I).
-    ceiling_tracks_input = [
-        {"filename": r["filename"], "lufs": r["lufs"]}
-        for r in verify_results
-    ]
-    ceiling_result = _ceiling_guard_compute_overshoots(ceiling_tracks_input)
+    if result := await _album_stages._stage_ceiling_guard(
+        ctx, _compute_overshoots=_ceiling_guard_compute_overshoots
+    ):
+        return result
 
-    ceiling_stage: dict[str, Any] = {
-        "status":       "pass",
-        "action":       "no_op",
-        "median_lufs":  None if ceiling_result["median_lufs"] is None
-                        else round(ceiling_result["median_lufs"], 2),
-        "threshold_lu": None if ceiling_result["threshold_lu"] is None
-                        else round(ceiling_result["threshold_lu"], 2),
-        "tracks":       ceiling_result["tracks"],
-        "pulled_down":  [],
-    }
-
-    halt_rows = [r for r in ceiling_result["tracks"] if r["classification"] == "halt"]
-    if halt_rows:
-        ceiling_stage["status"] = "fail"
-        ceiling_stage["action"] = "halt"
-        stages["ceiling_guard"] = ceiling_stage
-        return _safe_json({
-            "album_slug":      album_slug,
-            "stage_reached":   "ceiling_guard",
-            "stages":          stages,
-            "settings":        settings,
-            "warnings":        warnings,
-            "failed_stage":    "ceiling_guard",
-            "failure_detail": {
-                "reason": (
-                    f"Album-ceiling guard: "
-                    f"{len(halt_rows)} track(s) with overshoot > 0.5 LU. "
-                    f"Coherence correction (step 6) did not converge on LUFS-I; "
-                    f"re-run master_album with coherence corrections or adjust "
-                    f"per-track mastering."
-                ),
-                "tracks": [
-                    {
-                        "filename":      r["filename"],
-                        "lufs":          round(r["lufs"], 2),
-                        "overshoot_lu":  round(r["overshoot_lu"], 2),
-                    }
-                    for r in halt_rows
-                ],
-                "threshold_lu": ceiling_stage["threshold_lu"],
-                "median_lufs":  ceiling_stage["median_lufs"],
-            },
-        })
-
-    correctable_rows = [
-        r for r in ceiling_result["tracks"]
-        if r["classification"] == "correctable"
-    ]
-    if correctable_rows:
-        # Apply scalar pull-down in-place to each correctable track, then
-        # re-analyze so downstream stages (mastering samples, post-QC,
-        # archival, signature persist) see accurate numbers.
-        # lambda binds kwargs because run_in_executor only forwards positional
-        # args and apply_pull_down_db uses keyword-only params.
-        output_bits = int(targets.get("output_bits") or 24)
-        pull_down_errors: list[str] = []
-        pulled_files: list[str] = []
-        for row in correctable_rows:
-            target_path = output_dir / row["filename"]
-            try:
-                await loop.run_in_executor(
-                    None,
-                    functools.partial(
-                        apply_pull_down_db,
-                        target_path,
-                        gain_db=row["pull_down_db"],
-                        output_bits=output_bits,
-                    ),
-                )
-                pulled_files.append(row["filename"])
-            except CeilingGuardError as exc:
-                pull_down_errors.append(f"{row['filename']}: {exc}")
-
-        if pull_down_errors:
-            ceiling_stage["status"] = "warn"
-            ceiling_stage["pull_down_errors"] = pull_down_errors
-            warnings.append(
-                "Ceiling guard: pull-down failed for "
-                f"{len(pull_down_errors)} track(s); see stage output"
-            )
-        ceiling_stage["action"] = "pull_down"
-        ceiling_stage["pulled_down"] = pulled_files
-
-        # Patch verify_results so stages["verification"], mastering samples
-        # (Stage 5.5), post-QC (Stage 6), and archival (Stage 6.5) see
-        # post-pull-down numbers. Note: Stage 7.5 signature persistence uses
-        # analysis_results (pre-master), not verify_results, by design.
-        filename_to_index = {
-            r["filename"]: i for i, r in enumerate(verify_results)
-        }
-        for fname in pulled_files:
-            wav_path = output_dir / fname
-            fresh = await loop.run_in_executor(
-                None, analyze_track, str(wav_path),
-            )
-            idx = filename_to_index.get(fname)
-            if idx is not None:
-                verify_results[idx] = fresh
-
-        verify_lufs = [r["lufs"] for r in verify_results]
-        stages["verification"]["avg_lufs"] = round(float(np.mean(verify_lufs)), 1)
-        stages["verification"]["lufs_range"] = round(
-            float(max(verify_lufs) - min(verify_lufs)), 2
-        )
-
-    stages["ceiling_guard"] = ceiling_stage
+    verify_results = ctx.verify_results
 
     # --- Stage 5.5: Mastering samples (codec preview + mono fold-down QC) ---
     # Issue #296. Writes .aac.m4a and .MONO_FOLD.md sidecars to the
